+++ HISTORY.rst
@@ -16,6 +16,10 @@ Unreleased
 
 **Bugfixes**
 
+* Change the type of model for the pharmacy reception
+* Fix a bug with the name_search of product.product
+* Fix a bug with the wizard to "put in pack"
+
 **Build**
 
 * ALCYN-1756: Pin pending-merges to prevent unwanted changes